# Conflicts:
#	.gitignore
#	InstColorization.ipynb
#	LICENSE
#	README.md
#	README_TRAIN.md
#	calculate_metrics.py
#	checkpoints/.gitignore
#	data/__init__.py
#	data/aligned_dataset.py
#	data/base_data_loader.py
#	data/base_dataset.py
#	data/color_dataset.py
#	data/image_folder.py
#	data/single_dataset.py
#	download.py
#	env.yml
#	example/574100c6e31a1dcd096476eb2d632b3c.jpg
#	example/8b01a894fb91025a1dc77611838e9d13.jpg
#	example/925f179ba736e15e89bfd6d88e0bba56.jpg
#	example/e9b6826aa623549ec77bbc0275002779.jpg
#	fusion_dataset.py
#	image_util.py
#	imgs/teaser.png
#	inference_bbox.py
#	models/__init__.py
#	models/base_model.py
#	models/fusion_model.py
#	models/networks.py
#	models/train_model.py
#	options/__init__.py
#	options/base_options.py
#	options/train_options.py
#	scripts/download_model.sh
#	scripts/install.sh
#	scripts/prepare_cocostuff.sh
#	scripts/prepare_train_box.sh
#	scripts/test_mask.sh
#	scripts/train.sh
#	test_fusion.py
#	train.py
#	util/__init__.py
#	util/get_data.py
#	util/html.py
#	util/image_pool.py
#	util/util.py
#	util/visualizer.py

+++ inference_bbox.py
@@ -23,6 +23,7 @@ cfg = get_cfg()
 cfg.merge_from_file(model_zoo.get_config_file("COCO-InstanceSegmentation/mask_rcnn_X_101_32x8d_FPN_3x.yaml"))
 cfg.MODEL.ROI_HEADS.SCORE_THRESH_TEST = 0.7
 cfg.MODEL.WEIGHTS = model_zoo.get_checkpoint_url("COCO-InstanceSegmentation/mask_rcnn_X_101_32x8d_FPN_3x.yaml")
+# cfg.MODEL.DEVICE = "cpu"
 predictor = DefaultPredictor(cfg)
 
 parser = ArgumentParser()
@@ -38,16 +39,20 @@ if os.path.isdir(output_npz_dir) is False:
     os.makedirs(output_npz_dir)
 
 for image_path in tqdm(image_list):
-    img = cv2.imread(join(input_dir, image_path))
-    lab_image = cv2.cvtColor(img, cv2.COLOR_BGR2LAB)
-    l_channel, a_channel, b_channel = cv2.split(lab_image)
-    l_stack = np.stack([l_channel, l_channel, l_channel], axis=2)
-    outputs = predictor(l_stack)
     save_path = join(output_npz_dir, image_path.split('.')[0])
-    pred_bbox = outputs["instances"].pred_boxes.to(torch.device('cpu')).tensor.numpy()
-    pred_scores = outputs["instances"].scores.cpu().data.numpy()
-    if args.filter_no_obj is True and pred_bbox.shape[0] == 0:
-        print('delete {0}'.format(image_path))
-        os.remove(join(input_dir, image_path))
-        continue
-    np.savez(save_path, bbox = pred_bbox, scores = pred_scores)
+    if os.path.isfile(save_path):
+        print(f'{image_path} processed')
+    else:
+        img = cv2.imread(join(input_dir, image_path))
+        lab_image = cv2.cvtColor(img, cv2.COLOR_BGR2LAB)
+        l_channel, a_channel, b_channel = cv2.split(lab_image)
+        l_stack = np.stack([l_channel, l_channel, l_channel], axis=2)
+        outputs = predictor(l_stack)
+        # save_path = join(output_npz_dir, image_path.split('.')[0])
+        pred_bbox = outputs["instances"].pred_boxes.to(torch.device('cpu')).tensor.numpy()
+        pred_scores = outputs["instances"].scores.cpu().data.numpy()
+        if args.filter_no_obj is True and pred_bbox.shape[0] == 0:
+            print('delete {0}'.format(image_path))
+            os.remove(join(input_dir, image_path))
+            continue
+        np.savez(save_path, bbox = pred_bbox, scores = pred_scores)
